spotchecks (2)
