+++ test/integration/Jamfile
@@ -30,6 +30,7 @@ run
         src/async_callback.cpp
         src/async_coroutines.cpp
         src/async_coroutinescpp20.cpp
+        src/spotchecks_helpers.cpp
 
         # Actual tests
         test/spotchecks.cpp
